Introdotto databasewriter
