--- app/src/main/java/com/unison/appartment/database/Database.java
@@ -1,8 +1,0 @@
-package com.unison.appartment.database;
-
-import com.unison.appartment.model.User;
-
-public interface Database {
-    void writeUser(final User newUser, final DatabaseListener listener, final String uid);
-    void retrieveUser(final String uid, final DatabaseListener listener);
-}
