+++ app/src/main/java/com/unison/appartment/database/DatabaseWriter.java
@@ -1,0 +1,7 @@
+package com.unison.appartment.database;
+
+import com.unison.appartment.model.User;
+
+public interface DatabaseWriter {
+    void writeUser(final User newUser, final String uid, final DatabaseWriterListener listener);
+}
