+++ app/src/main/java/com/unison/appartment/database/DatabaseWriterListener.java
@@ -2,13 +2,7 @@ package com.unison.appartment.database;
 
 import com.google.firebase.database.DatabaseError;
 
-public interface DatabaseListener {
-
-    void onReadSuccess(Object object);
-    void onReadEmpty();
-    void onReadCancelled(DatabaseError databaseError);
-
+public interface DatabaseWriterListener {
     void onWriteSuccess();
-
     void onWriteFail(Exception exception);
 }
